+++ src/services/dispatcher/BlockChainDispatcher.ts
@@ -7,6 +7,10 @@ import {BlockchainRepository} from '../../repositories/BlockchainRepository';
 import Settings from '../../types/Settings';
 import {ChainsIds} from '../../types/ChainsIds';
 import Blockchain from '../../lib/Blockchain';
+import {AvalancheBlockDispatcher} from './AvalancheBlockDispatcher';
+import {PolygonBlockDispatcher} from './PolygonBlockDispatcher';
+import {ArbitrumBlockDispatcher} from './ArbitrumBlockDispatcher';
+import {EthereumBlockDispatcher} from './EthereumBlockDispatcher';
 
 export type BlockChainDispatcherProps = {
   chainId: ChainsIds;
@@ -20,9 +24,19 @@ export class BlockChainDispatcher {
   @inject(BlockchainRepository) blockchainRepository!: BlockchainRepository;
   @inject('Settings') private readonly settings!: Settings;
 
-  constructor(@inject(BSCBlockDispatcher) bscBlockDispatcher: BSCBlockDispatcher) {
+  constructor(
+    @inject(BSCBlockDispatcher) bscBlockDispatcher: BSCBlockDispatcher,
+    @inject(AvalancheBlockDispatcher) avalancheBlockDispatcher: AvalancheBlockDispatcher,
+    @inject(PolygonBlockDispatcher) polygonBlockDispatcher: PolygonBlockDispatcher,
+    @inject(ArbitrumBlockDispatcher) arbitrumBlockDispatcher: ArbitrumBlockDispatcher,
+    @inject(EthereumBlockDispatcher) ethereumBlockDispatcher: EthereumBlockDispatcher,
+  ) {
     this.dispatchers = {
       bsc: bscBlockDispatcher,
+      avax: avalancheBlockDispatcher,
+      ethereum: ethereumBlockDispatcher,
+      polygon: polygonBlockDispatcher,
+      arbitrum: arbitrumBlockDispatcher,
     };
   }