+++ test/services/multichain/MultiChainStatusProcessor.test.ts
@@ -3,7 +3,6 @@ import {expect} from 'chai';
 
 import {getTestContainer} from '../../helpers/getTestContainer';
 import {MultiChainStatusProcessor} from '../../../src/services/multiChain/MultiChainStatusProcessor';
-import {ChainsIds} from '../../../src/types/ChainsIds';
 import {ChainStatusWithAddress} from '../../../src/types/ChainStatus';
 import {chainStatusWithAddressFactory, chainStatusFactory} from '../../mocks/factories/chainStatusFactory';
 
@@ -11,7 +10,6 @@ describe('MultiChainStatusProcessor', () => {
   let chainStatusWithAddress: ChainStatusWithAddress[];
   let multiChainStatusProcessor: MultiChainStatusProcessor;
   const bscChainStatus = chainStatusWithAddressFactory.build();
-  const chainsIds = Object.values(ChainsIds);
 
   chainStatusWithAddress = [
     bscChainStatus,
@@ -31,10 +29,7 @@ describe('MultiChainStatusProcessor', () => {
   describe('when all chains can mint', () => {
     it('returns the chainsIdsReadyForBlock for all chains', async () => {
       const result = await multiChainStatusProcessor.apply(chainStatusWithAddress, 10);
-
-      chainsIds.forEach((chain, i) => {
-        expect(result.chainsStatuses[i]).to.includes({chainId: chain});
-      });
+      expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
 
       expect(result.chainsIdsReadyForBlock).to.deep.equal(['bsc', 'avax']);
     });
@@ -60,11 +55,7 @@ describe('MultiChainStatusProcessor', () => {
 
     it('returns the chainsIdsReadyForBlock for the one chain that can mint', async () => {
       const result = await multiChainStatusProcessor.apply(chainStatusWithAddress, 10);
-
-      chainsIds.forEach((chain, i) => {
-        expect(result.chainsStatuses[i]).to.includes({chainId: chain});
-      });
-
+      expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
       expect(result.chainsIdsReadyForBlock).to.deep.equal(['bsc']);
     });
   });